+++ docs/source/changelog.rst
@@ -7,6 +7,9 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ==========
 
+* Overloaded ``VaultCustomQueryCriteria`` constructor to allow clients to specify the Hibernate ``Entity`` class manually.
+  This is required in case a ``CriteriaExpression`` references a field from a ``MappedSuperclass``.
+
 * Added program line argument ``on-unknown-config-keys`` to allow specifying behaviour on unknown node configuration property keys.
   Values are: [FAIL, WARN, IGNORE], default to FAIL if unspecified.